% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchGSEA.R
\name{BatchGSEA}
\alias{BatchGSEA}
\title{Performs GSEA for all Hotgenes contrasts}
\usage{
BatchGSEA(
  m_df = NULL,
  HotgenesObj = NULL,
  nTop = 10,
  minSize = 15,
  maxSize = 1000,
  nperm = 1e+05,
  nproc = 4
)
}
\arguments{
\item{m_df}{GeneSet generated using msigdbr function}

\item{HotgenesObj}{R object generated by DEseq2_export function}

\item{nTop}{number of pathways to return}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{nperm}{Number of permutations to do. Minimial possible nominal p-value is about 1/nperm}

\item{nproc}{If not equal to zero sets BPPARAM to use nproc workers (default = 0).}
}
\value{
HotgenesObj appended with GSEA for each contrast
}
\description{
Performs GSEA for all Hotgenes contrasts
}
\examples{
Example_Hotgenes_dir<-system.file("extdata",
"Example_Hotgenes.Rdata",
package = "Hotgenes", mustWork = TRUE)
load(Example_Hotgenes_dir)
library(msigdbr)
m_df = msigdbr(species = "Homo sapiens", category = "C5", subcategory = "BP")
qbat<-BatchGSEA(m_df= m_df, HotgenesObj=Example_Hotgenes)
summary(qbat)
qbat$OuputGSEA$Hrs2$g
}
