% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factominer_tables.R
\name{Categorical_Table}
\alias{Categorical_Table}
\title{Generates a table of Categorical variables from FactoMineR data}
\usage{
Categorical_Table(res.hcpc = NULL, Round_To = 2, Format = TRUE)
}
\arguments{
\item{res.hcpc}{R object generated by HCPC function
from FactoMineR}

\item{Round_To}{numeric value indicating digit for rounding
default = 2}

\item{Format}{Logical, if TRUE (default), output formated by
group.}
}
\value{
data.frame with HCPC results
}
\description{
Generates a table of Categorical variables from FactoMineR data
}
\examples{
Example_Hotgenes_dir<-system.file("extdata",
"Example_Hotgenes.Rdata",
package = "Hotgenes", mustWork = TRUE)
load(Example_Hotgenes_dir)
Results<-FactoWrapper(Example_Hotgenes)
Categorical_Table(Results$res.hcpc)
}
