% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEseq2_export.R
\name{DEseq2_export}
\alias{DEseq2_export}
\title{Prepares DESeq2 object for visualization}
\usage{
DEseq2_export(DEseq2_object = NULL, padj_cut = 0.1)
}
\arguments{
\item{DEseq2_object}{the output object generated by DESeq2}

\item{padj_cut}{the adjusted pvalue cut off for differential
expression.  Default = 0.1}
}
\description{
Prepares DESeq2 object for visualization
}
