% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FactoWrapper.R
\name{FactoWrapper}
\alias{FactoWrapper}
\title{FactoMiner wrapper used by Shiny_DE_Viz for PCA/HCPC of samples and genes}
\usage{
FactoWrapper(
  Output_DEseq2 = NULL,
  biplot = TRUE,
  readouts = 1,
  Top_var = 10,
  ellipse.level = 0.5,
  ellipse.alpha = 0.5,
  habillage_selection = "clust",
  label_sel = c("ind", "var"),
  pointsize = 1,
  labelsize = 1,
  selected_contrast = 1
)
}
\arguments{
\item{Output_DEseq2}{R object generated by DEseq2_export function}

\item{biplot}{Boolean value. If TRUE (Default), both individuals and variables
will be shown on the plot. If FALSE, only individuals will be plotted.}

\item{readouts}{either "rlog" (default) or "vsd" indicating the variance
stabilization method preferred for visualization.}

\item{Top_var}{a numeric value indicating the number of top contributing
genes to show on the biplot.}

\item{ellipse.level}{numeric value for ellipse size, from 0 to 1.0}

\item{ellipse.alpha}{numeric value for ellipse transparency, from 0 to 1.0.}

\item{habillage_selection}{string indicating the grouping to be shown.
Default is "clust", which represents the HCPC ward clusters.}

\item{label_sel}{String indicating if individuals "ind" and variables
"var" should be labeled on the biplot.}

\item{pointsize}{numeric value for point size.}

\item{labelsize}{numeric value for label size.}

\item{selected_contrast}{string indicating the DE contrast to be explored.}
}
\value{
FactoMiner plot showing HCPC of genes, samples, and conditions
}
\description{
FactoMiner wrapper used by Shiny_DE_Viz for PCA/HCPC of samples and genes
}
\examples{
Example_Hotgenes_dir<-system.file("extdata",
"Example_Hotgenes.Rdata",
package = "Hotgenes", mustWork = TRUE)
load(Example_Hotgenes_dir)
FactoWrapper(Example_Hotgenes)
}
