% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLocationModel.R
\name{generateLocationModel}
\alias{generateLocationModel}
\title{\code{generateLocationModel}}
\usage{
generateLocationModel(startBase, endBase, scaling = 1000)
}
\arguments{
\item{startBase}{The starting base pair number}

\item{endBase}{The ending base pair number}

\item{scaling}{scaling factor in how many base pairs should be represented per index (ex. 1000) to reduce computational complexity when simulating physics/plotting;
Should match value used in generateStrandModel}
}
\value{
A sequence containing the location model.
}
\description{
\code{generateLocationModel}
}
