% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateStrandModel.R
\name{generateStrandModels}
\alias{generateStrandModels}
\title{\code{generateStrandModels}}
\usage{
generateStrandModels(startBase, endBase, fc, strand)
}
\arguments{
\item{startBase}{The starting base pair number}

\item{endBase}{The ending base pair number}

\item{fc}{The featurecounts object from rsubread after counting your alignment}

\item{strand}{+ or - strand to generate models for}
}
\value{
A data frame containing the strand model.
}
\description{
\code{generateStrandModels}
}
