% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSet_Exprs.R
\name{GeneSet_Exprs}
\alias{GeneSet_Exprs}
\title{Exports Expression data for a GeneSet of interest}
\usage{
GeneSet_Exprs(
  GSEA_Hotgenes = NULL,
  normalizedData = 1,
  contrastID = 1,
  pathwayID = 1
)
}
\arguments{
\item{GSEA_Hotgenes}{HotgenesObj appended by BatchGSEA function}

\item{normalizedData}{indicates the normalized data in
HotgenesObj to return.}

\item{contrastID}{contrast selection}

\item{pathwayID}{String or numeric index indicating the desired pathway}
}
\value{
Expresion data
}
\description{
Exports Expression data for a GeneSet of interest
}
\examples{
Example_Hotgenes_dir<-system.file("extdata",
"Example_Hotgenes.Rdata",
package = "Hotgenes", mustWork = TRUE)
load(Example_Hotgenes_dir)
library(msigdbr)
m_df = msigdbr(species = "Homo sapiens", category = "C5", subcategory = "BP")
qbat<-BatchGSEA(m_df= m_df, HotgenesObj=Example_Hotgenes)
expsD<-GeneSet_Exprs(GSEA_Hotgenes=qbat)
library(pheatmap)
pheatmap(expsD$E,scale = "row",
main=expsD$Main)
}
