% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTA_Annotations.R
\name{get_anno}
\alias{get_anno}
\title{Prepares gets annotation from FASTA file}
\usage{
get_anno(trans_fasta)
}
\arguments{
\item{trans_fasta}{object generated by readDNAStringSet
from Biostrings package}
}
\description{
Prepares gets annotation from FASTA file
}
