% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hotgenes2GeneSet.R
\name{Hotgenes2GeneSet}
\alias{Hotgenes2GeneSet}
\title{Generates a custom GeneSet from a Hotgenes object}
\usage{
Hotgenes2GeneSet(Hotgenes_obj = NULL)
}
\arguments{
\item{Hotgenes_obj}{R object generated by DEseq2_export function}
}
\value{
data.frame with "gene_symbol" and "gs_name".
GeneSet contains genes that are either Enriched or Depleted by the
condition (based on logfoldchange).
}
\description{
Generates a custom GeneSet from a Hotgenes object
}
\examples{
Example_Hotgenes_dir<-system.file("extdata",
"Example_Hotgenes.Rdata",
package = "Hotgenes", mustWork = TRUE)
load(Example_Hotgenes_dir)
Custom_GeneSet<-Hotgenes2GeneSet(Example_Hotgenes)
head(Custom_GeneSet)
}
