% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hotgenes2csv.R
\name{hotgenes_csv}
\alias{hotgenes_csv}
\title{Exports hotgenes object to csv files}
\usage{
hotgenes_csv(
  hotgenes = NULL,
  Exps_Out = TRUE,
  readouts = 1,
  sep = "\\t",
  dir_out = NULL
)
}
\arguments{
\item{hotgenes}{object created by DEseq2_export or Limma_export
functions.}

\item{Exps_Out}{logical, if TRUE (default), available expression data
is also exported to csv file.}

\item{readouts}{integer indicated the desired normalized data to be 
written to csv file. If 1 (default), the first method is exported}

\item{sep}{field separator string}

\item{dir_out}{string providing the desired directory for export. 
If not provided, the working directory will be used.}
}
\value{
multiple csv files
}
\description{
Exports hotgenes object to csv files
}
\examples{
Example_Hotgenes_dir<-system.file("extdata",
"Example_Hotgenes.Rdata",
package = "Hotgenes", mustWork = TRUE)
load(Example_Hotgenes_dir)
hotgenes_csv(Example_Hotgenes)
}
