% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limma_Adj_pval.R
\name{Limma_Adj_pval}
\alias{Limma_Adj_pval}
\title{Returns adjusted pvalue from decideTests function, instead of -1, 0 or 1}
\usage{
Limma_Adj_pval(
  object,
  method = "separate",
  adjust.method = "BH",
  p.value = 0.05,
  lfc = 0
)
}
\arguments{
\item{object}{a numeric matrix of p-values or an \code{MArrayLM} object from which p-values and t-statistics can be extracted.}

\item{method}{character string specifying how genes and contrasts are to be combined in the multiple testing scheme.  Choices are \code{"separate"}, \code{"global"}, \code{"hierarchical"} or \code{"nestedF"}.}

\item{adjust.method}{character string specifying p-value adjustment method.  Possible values are \code{"none"}, \code{"BH"}, \code{"fdr"} (equivalent to \code{"BH"}), \code{"BY"} and \code{"holm"}. See \code{\link[stats]{p.adjust}} for details.}

\item{p.value}{numeric value between 0 and 1 giving the required family-wise error rate or false discovery rate.}

\item{lfc}{numeric, minimum absolute log2-fold-change required.}
}
\description{
Returns adjusted pvalue from decideTests function, instead of -1, 0 or 1
}
