% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatedMap.R
\name{plotHeatedMap}
\alias{plotHeatedMap}
\title{\code{plotHeatedMap}}
\usage{
plotHeatedMap(expression, locations)
}
\arguments{
\item{expression}{Input dataframe for adjusted RNA-seq values. the dataframe should be organized such that RNA-seq values run horizontally;
this means that after loading column vectors the data frame should be transposed with t(). In particular if you get "figure margins too large"
you are probably not transposing either this or the other parameter correctly.}

\item{locations}{Input dataframe for the base pair number cooresponding with expression. This should have the same dimensions and organization as expression.}
}
\description{
\code{plotHeatedMap} is used to create a plot of a heated map of the specified data frame.
}
