% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Factominer_tables.R
\name{Quanti_Table}
\alias{Quanti_Table}
\title{Generates a table of quantitative variables from FactoMineR data}
\usage{
Quanti_Table(res.hcpc = NULL, Round_To = 2)
}
\arguments{
\item{res.hcpc}{R object generated by HCPC function
from FactoMineR}

\item{Round_To}{numeric value indicating digit for rounding
default = 2}
}
\value{
data.frame with HCPC results
}
\description{
Generates a table of quantitative variables from FactoMineR data
}
\examples{
Example_Hotgenes_dir<-system.file("extdata",
"Example_Hotgenes.Rdata",
package = "Hotgenes", mustWork = TRUE)
load(Example_Hotgenes_dir)
Results<-FactoWrapper(Example_Hotgenes)
Quanti_Table(Results$res.hcpc)
}
