% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateStrandModel.R
\name{rebuildStrandModel}
\alias{rebuildStrandModel}
\title{Rebuild a Strand Model}
\usage{
rebuildStrandModel(strandModel, newStartBase, newEndBase, newScaling,
  startBase, endBase, scaling)
}
\arguments{
\item{strandModel}{a strand model that is a superset of the new strand model with tighter scaling}

\item{newStartBase}{an integer depicting the desired start base pair number.}

\item{newEndBase}{an integer depicting the desired end base pair number.}

\item{newScaling}{scaling factor in how many base pairs should be represented per index. Should be larger (eg. 1000000) than old scaling factor.
IMPORTANT: newScaling is assumed to be divisible by scaling.}

\item{startBase}{an integer depicting the old strand model's starting base pair number. Should be the beginning of the chromosome to be rendered (eg. 1).}

\item{endBase}{an integer depicting the old strand model's starting base pair number. Should be the end of the chromosome to be rendered (eg. 3000000).}

\item{scaling}{scaling factor in how many base pairs should be represented per index in the existing strand model. Should be smaller than any rebuilt choice (eg. 1)}
}
\value{
a strand model with the new start-end base and/or more relaxed scaling
}
\description{
This function is cheaper than rebuilding a brand new strand model from scratch. These savings are realized by using parts of the old strand model.
}
\examples{
\dontrun{
   sm <- generateStrandModel(startBase=1, endBase=195471971,
    fc=musCh1fc, chr="chr1", strand="-", scaling=1)
   sm1 <- rebuildStrandModel(strandModel=sm, newStartBase=1000,
    newEndBase=2000, newScaling=100000, startBase=1, endBase=195471971, scaling=1)
}

}
