% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHotgenesShiny.R
\name{runHotgenes}
\alias{runHotgenes}
\title{Launches the shiny app for hotgenes}
\usage{
runHotgenes()
}
\value{
No return value, but starts a shiny server.
}
\description{
A function that launches the shiny app for hotgenes.
}
\examples{
\dontrun{

}

}
