% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shiny_DE_viz.R
\name{Shiny_DE_viz}
\alias{Shiny_DE_viz}
\title{R shiny app for visualizing DE analysis}
\usage{
Shiny_DE_viz(DEseq2_export = NULL)
}
\arguments{
\item{DEseq2_export}{R object generated by DEseq2_export function}
}
\value{
a shiny object for network visualization.

a shiny object for visualization.
}
\description{
R shiny app for visualizing DE analysis
}
\examples{
Example_Hotgenes_dir<-system.file("extdata",
"Example_Hotgenes.Rdata",
package = "Hotgenes", mustWork = TRUE)
load(Example_Hotgenes_dir)
if(interactive()){
Shiny_DE_viz(Example_Hotgenes)}
}
