% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateHeatSpread.R
\name{simulateHeatSpread}
\alias{simulateHeatSpread}
\title{\code{simulateHeatSpread}}
\usage{
simulateHeatSpread(.strandModel, conductivity, iterations)
}
\arguments{
\item{.strandModel}{Strand model from generateStrandModel}

\item{conductivity}{numeric between 0-1: 1 implies perfect equalization of heat between two adjacent cells in 1 iteration, 0 implies no heat transfer}

\item{iterations}{integer number of cycles to simulate: should be 1 or greater}
}
\value{
simulated strand model
}
\description{
Simulates heat spread through conductance on (a) strand model(s). Kind of looks like anti-aliasing. Makes smaller expressed exons more visible.
}
\examples{
\dontrun{
sm <- generateStrandModels(1, 195471971, musCh1fc, "chr1", "-", 100000)
smh <- simulateHeatSpread(sm, 0.001, 1000)
smh1 <- simulateHeatSpread(sm, 1, 100)
}

}
