% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEseq2_export.R
\name{DEseq2_export}
\alias{DEseq2_export}
\title{Prepares DESeq2 object for visualization}
\usage{
DEseq2_export(
  DEseq2_object = NULL,
  lfcShrink_type = "normal",
  rlog_export = TRUE,
  padj_cut = 0.1
)
}
\arguments{
\item{DEseq2_object}{the output object generated by DESeq2}

\item{lfcShrink_type}{"normal" (Default) is the original DESeq2 shrinkage
estimator. Other options is "apeglm".}

\item{rlog_export}{Logical value, if TRUE (default), data
normalized by rlog will be exported. If FALSE, only 
varianceStabilizingTransformation normalization will be used.}

\item{padj_cut}{the adjusted pvalue cut off for differential
expression.  Default = 0.1}
}
\value{
list of R objects needed for Shiny_DE_viz function
}
\description{
Prepares DESeq2 object for visualization
}
\examples{
dds_con_dir<-system.file("extdata",
"dds_con.Rdata", package = "Hotgenes", mustWork = TRUE)
load(dds_con_dir)
Example_Hotgenes<-DEseq2_export(DEseq2_object = dds_con, padj_cut = 0.1)
Example_Hotgenes$Exported_plots$transformation_plots
Example_Hotgenes$Exported_plots$boxplot_rlog
Example_Hotgenes$Exported_plots$boxplot_vsd
}
\seealso{
\code{\link[DESeq2]{lfcShrink}}
}
