% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateStrandModel.R
\name{generateStrandModels}
\alias{generateStrandModels}
\alias{generateStrandModel}
\title{\code{generateStrandModels}}
\usage{
generateStrandModels(startBase, endBase, fc, chr, strand, scaling = 1000)

generateStrandModel(startBase, endBase, fc, chr, strand, scaling = 1000)
}
\arguments{
\item{startBase}{The starting base pair number}

\item{endBase}{The ending base pair number}

\item{fc}{The featurecounts object from rsubread after counting your alignment}

\item{chr}{Chromosome; Should match Chr in fc}

\item{strand}{+ or - strand to generate models for}

\item{scaling}{scaling factor in how many base pairs should be represented per index (ex. 1000) to reduce computational complexity when simulating physics/plotting}
}
\value{
A matrix containing the strand model.
}
\description{
\code{generateStrandModels}
}
