% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateStrandModel.R
\name{generateStrandModels}
\alias{generateStrandModels}
\alias{generateStrandModel}
\title{\code{generateStrandModel}

Creates a model of a DNA strand in the form of a matrix (data frames are too slow) for further processing/plotting.}
\usage{
generateStrandModel(startBase, endBase, fc, chr, strand, scaling = 1000)
}
\arguments{
\item{startBase}{The starting base pair number}

\item{endBase}{The ending base pair number}

\item{fc}{The featurecounts object from rsubread after counting your alignment}

\item{chr}{Chromosome; Should match Chr in fc}

\item{strand}{+ or - strand to generate models for}

\item{scaling}{scaling factor in how many base pairs should be represented per index (ex. 1000) to reduce computational complexity when simulating physics/plotting}
}
\value{
A matrix containing the strand model.
}
\description{
\code{generateStrandModel}

Creates a model of a DNA strand in the form of a matrix (data frames are too slow) for further processing/plotting.
}
\examples{
\dontrun{
   #featureCounts() is from RSubread package; only run this if that is installed
   fc <- featureCounts(files=bamFile, annot.ext=annotation, useMetaFeatures=FALSE)
   #replace fc with the one loaded from featureCounts if you are using your own data
   sm <- generateStrandModel(startBase=1, endBase=195471971,
    fc=musCh1fc, chr="chr1", strand="-", scaling=100000)
   sm1 <- generateStrandModel(1, 195471971, musCh1fc, "chr1", "+")
   sm2 <- generateStrandModel(3000000, 3217000, musCh1fc, "chr1", "-")
}

}
