% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Limma_export.R
\name{Limma_export}
\alias{Limma_export}
\title{Prepares Limma object for visualization}
\usage{
Limma_export(limmafit = NULL, design_data = NULL,
  Expression_dat = NULL, method = "global", adjust.method = "BH",
  p.value = 0.1)
}
\arguments{
\item{limmafit}{the fit generated by limma}

\item{design_data}{data.frame containing Experimental design}

\item{method}{character string specifying how genes and contrasts are to be combined in the multiple testing scheme.  Choices are \code{"separate"}, \code{"global"}, \code{"hierarchical"} or \code{"nestedF"}.}

\item{adjust.method}{character string specifying p-value adjustment method.  Possible values are \code{"none"}, \code{"BH"}, \code{"fdr"} (equivalent to \code{"BH"}), \code{"BY"} and \code{"holm"}. See \code{\link[stats]{p.adjust}} for details.}

\item{p.value}{numeric value between 0 and 1 giving the required family-wise error rate or false discovery rate.}
}
\value{
list of R objects needed for Shiny_DE_viz function
}
\description{
Prepares Limma object for visualization
}
\examples{
Example_Hotgenes_dir<-system.file("extdata",
"Example_Hotgenes.Rdata",
package = "Hotgenes", mustWork = TRUE)
load(Example_Hotgenes_dir)

library(limma)

exp<-Example_Hotgenes$Normalized_Expression$rld
design_m<-Example_Hotgenes$design_data

design_matrix <- model.matrix(~sh*Hrs+Bio_Rep,   
data = design_m)

aw <- arrayWeights(exp, design_matrix)

fit <- lmFit(exp, design=design_matrix, weights = aw)
fit <- eBayes(fit, robust = TRUE) 

L_out<-Limma_export(Expression_dat = exp, 
design_data = design_m, 
limmafit = fit)
summary(L_out$Output_lists)
}
