% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vplot.R
\name{Vplot}
\alias{Vplot}
\title{Wrapper for EnhancedVolcano}
\usage{
Vplot(Hotgenes_input = NULL, pval_cut = 0, FCcutoff = 0,
  stuffer = 0.5, contrast = 1)
}
\arguments{
\item{Hotgenes_input}{R object generated by DEseq2_export or 
Limma_export functions}

\item{pval_cut}{numeric indicating pvalue cut off}

\item{FCcutoff}{numeric indicating logfoldchange cut off}

\item{stuffer}{is a positive number added or subtracted from pvalue 
or fold change values, used for x and y limits.}

\item{contrast}{name or numeric value to retrieve differential 
expression data. See names(Hotgenes_input$Output_DE).}
}
\description{
Wrapper for EnhancedVolcano
}
\examples{
Example_Hotgenes_dir<-system.file("extdata",
"Example_Hotgenes.Rdata",
package = "Hotgenes", mustWork = TRUE)
load(Example_Hotgenes_dir)
names(Example_Hotgenes$Output_DE)
Vplot(Example_Hotgenes, TopSig=10, contrast = 1)
}
